Changed configuration to point to local installables.
Former-commit-id: c4f1ae6c787cdee0ddafbf94d0fe9f4d63311ed7 [formerly 2cf59e6bebf49dcd605da7eb065e119877b11916] [formerly 00337b275d07c8f3b4fbdea8cbc811f8767113b3 [formerly ae44d9608dd6c6cb1a3d1e284ebe59bfeb8fedfa]] [formerly bf3291c27f3472f9c3e8b82bfb1a3772c5f0bb60] [formerly 31e7f4235df965a5846cbd58cefe31cbdb3fcbe6 [formerly dfbb53e65a379a5c5c47c6fa922abf415aa7e5fe]] [formerly 9b169e28528dac3d8540f2e2208c6d539b9ee337] [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7] [formerly 085f07f7710252e8f619fbd4c4e79f31849dcf7d [formerly 0b556b541c739847bbfea8feccd1702d6f9aea1a] [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly e99229da2259afb3aad028cbd1d37bcc4436a2e7 [formerly 3447b24c16a1b7b0bb1ec545edc68e35c741a3cf]]]]]]]
Former-commit-id: c4b22512e72c344b9bf29535a5a65ac098e35df2 [formerly 9803a5880b4979beb1f7adbdf61b13e20fedce72] [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4]] [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4]] [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4]] [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4] [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4 [formerly a65173e83f671368cef5b8101b9d8c5dde4f60be]]]]]]
Former-commit-id: e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly e17f1f68cff7547245e68fddd041a2ad43a956e6 [formerly 12a3dc5ff847e0c401db8186fb4ccdb6a031b9c4]]
Former-commit-id: 9502cc536e9b998fb996221309a746fe367029ae
Former-commit-id: 14696e1865b2711d12aa856da325b0bb43fa18c9 [formerly 86fff601d56267f62fa63ea5778b7cd746c7b727]
Former-commit-id: 35b59f128583e8e25bebd391832f731809a0af52
Former-commit-id: 828c80bb5b48f03cd7f8d96647dd25bf931b5a6b [formerly f3ab3dbe74049488f43bc5f92d3fbed9e0ec3af4]
Former-commit-id: fd1832188ed5d9ff8a5ac363b2bb1b5f03de1879

+++ scripts/sonar-setup/setup.sh
@@ -5,12 +5,27 @@
 mkdir /opt/sonar
 chmod 777 /opt/sonar
 cd /opt/sonar
-yum install -y unzip
-wget http://dist.sonar.codehaus.org/sonarqube-5.1.zip
+
+echo ...........................Installing Unzip...........................
+if [ -f "/vshare/base-images/jenkins/apache-maven-3.0.5-bin.tar.gz" ]; then
+	yum install -y /vshare/base-images/misc/unzip-6.0-13.el7.x86_64.rpm
+else
+	yum install -y unzip
+fi
+
+echo ...........................Installing Sonarqube...........................
+if [ -f "/vshare/base-images/sonarqube/sonarqube-5.1.zip" ]; then
+	cp /vshare/base-images/sonarqube/sonarqube-5.1.zip .
+else
+	wget http://dist.sonar.codehaus.org/sonarqube-5.1.zip
+fi
+
 chmod 777 sonarqube-5.1.zip
 unzip sonarqube-5.1.zip
 chmod 777 sonarqube-5.1
 cd sonarqube-5.1
+
+echo ...........................Changing Sonarqube configuration...........................
 cp /mnt/gluster/repo/sonar-setup/config/sonar.properties conf/
 chmod -R 777 conf/*
 cp /mnt/gluster/repo/sonar-setup/plugins/sonar-tanaguru-latest.jar extensions/plugins/
@@ -19,4 +34,6 @@ cp /mnt/gluster/repo/sonar-setup/config/sonar.sh /etc/init.d/sonar
 chkconfig --del sonar
 chmod 755 /etc/init.d/sonar
 chkconfig --add sonar
+
+echo ...........................Starting Sonarqube...........................
 service sonar start
